refactor: isolate PDF semantic flow validation
